decagon_voice/requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
gspread>=6.0.0
google-auth>=2.23.0
